backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3

